Improve EventListenerList Performance Bugzilla #255534
By separating the event listener types into individual lists and using a
map the startup performance could be greatly reduced. It is for small
number of EditParts (i.e., 2000) already twice as fast as the previous
implementation.
Furthermore iterating through the event listeners is simpler and most
probably also faster.

The work is based on the patch submitted by Iwan Birrer to the following
Bugzilla entry:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=255534