Allow Rulers with a Custom Dots Per Unit Setting
With this change the API of the RulerProvider class is extended with a
custom dots per unit setting. This allows clients to use our default
Ruler Figure adjusted to application specific resolutions and drawing
scales.

As part of this change the RulerFigure was cleaned by splitting the
paintFigure method in several methods for increased readability and
maintainability.

Bugzilla entry: https://bugs.eclipse.org/bugs/show_bug.cgi?id=105493