Restore ability to modify selection in GroupRequest within DeleteAction
The list returned by getSelectedEditParts() should be a view of the
internal selection, similar to getSelectedObjects(). Meaning callers are
able to modify by the selection by calling set(int,Object).

In order to reduce the chance of an undesirable ClassCastException, we
only check whether the first element in the list is an EditPart,
assuming that all remaining elements are as well. I.e. we don't expect a
"mixed" selection of different types. This is similar to what has
already been done in e.g. the DeleteAction or the MatchSizeAction.

This is a follow-up to 78c3fed3ced2c035c5a495a752523bb4b758a916.

Resolves https://github.com/eclipse-gef/gef-classic/issues/616