Disable testAntialias() test when running on Linux
Due to changes in SWT, the underlying Cairo surface of an SWT image is
no longer initialized with the system font options.

If a string is drawn without ever calling setTextAntialias(), the Cairo
default is used. If this method is called with SWT.DEFAULT, the system
default is used. Because those values are not guaranteed to be equal,
this may lead to visual artifacts.

See https://github.com/eclipse-gef/gef-classic/issues/696