Cleanup Draw2D PrintExample
If no default printer is defined, clicking on the "Print" button fails
with an SWT error. To avoid this (and to also avoid accidentally
printing to a real printer), a print dialog is opened to let the user
select the target.

This also replaces the MouseListener with a MouseAdapter, to get rid of
the empty listener methods.