Make DirectEditManager properly handle changed cell editor state
The DirectEditManager maintains an own dirty state to identify if a
command has to be executed on commit. In case it has an underlying cell
editor, this is when the cell editor submits its change. However, a
dirty state is already tracked by the underlying cell editor in that
case and the two are not necessarily consistent.
The DirectEditManager only updates its dirty state on a
`editorValueChanged` event from the cell editor, which is by definition
only sent on textual input changes. In case a selection is made, such as
in case of the ComboBoxCellEditor, no such `editorValueChanged` event is
submitted but only the `applyEditorValue` event is triggered. In
consequences, changes in ComboBoxCellEditors are currently not properly
converted into according commands by the DirectEditManager.

This change adapts the DirectEditManager to consider the dirty state of
the underlying cell editor in addition to a potentially set custom dirty
state on the DirectEditManager itself.