Separate cursor icon from its decoration for the shared cursors.
The way cursors are rendered on Windows is that regardless of the zoom
level, the stroke width of the cursor is always 1px wide. When using an
SVG, this would normally be achieved by setting the "non-scaling-stroke"
vector effect.

Sadly, this effect is not supported by the jSVG rasterizer that is used
by SWT and as a result, the stroke width is scaled proportionally to the
zoom level. I.e. 1px at 100%, 2px at 200% and so on.

To work around this limitation, the cursor is separated from the
remaining components and constructed on-the-fly, The complete cursor is
then a composite of the SVG and this generated cursor.