Make DirectEditManager more robust when brought down multiple times #757
When bringDown() is called while processing a "direct edit" command, an
IllegalArgumentException is thrown because the edit-part listener is
removed twice.

This is because the listener is null in the second round, which is not a
valid parameter to pass to the underlying EventListenerList of the edit
part.

An explicit null-check has been added to the DirectEditManager to avoid
this issue.