Implement a monitor-aware lightweight system to handle HiDPI scaling
This lightweight-system disables the native HiDPI scaling that is
normally done by SWT and instead scales the figures via a
ScalableLayeredPane.

The main purpose of this class is to reduce the noise that is introduced
by rounding, errors when dealing with fractional scaling.

To test this functionality, simply pass the MonitorAwareLightweighSystem
as an additional argument when constructing the FigureCanvas. See the
AutoScaleExample for reference.