Adapt API filters for removal of monitor-aware SWT classes
This is a follow-up to 50e27e3f2f36a1c1d161a219e3088e8ba6d4f1ed, this
time for the Zest plug-in. Because it also re-exports the SWT bundle, it
is also affected by the removal of the "MonitorAware" SWT classes [1].

This change suppresses those errors, as they are not relevant for
Zest.

[1] https://github.com/eclipse-platform/eclipse.platform.swt/pull/2349