Use auto-generated "disabled" icon when supported by SWT
SWT version 3.130.0 or newer is able to automatically generate a
"disabled" icon. If such a version is used, it is no longer necessary to
explicitly set it for e.g. an action.

In short: If SWT < 3.130 is used, both "enabled" and "disabled" icons
are set, otherwise only the "enabled" icon is set.

Closes https://github.com/eclipse-gef/gef-classic/issues/732