Avoid direct access to deprecated focusPart field in RulerComposite
If a "FocusIn" event is fired for the RulerComposite, the contents of
the RulerViewer should be focused, unless any other part has already
requested focus.

Rather than checking whether the deprecated field is checked, we can
check whether the FOCUS flag has been set for the focused edit part.
This flag is only set when setFocus(...) is called on the EditPartViewer
and should therefore be set exactly when the local variable is set.

Conversely, the focus should be cleared by an "FocusOut" event, unless
any other part has taken the focus.

Contributes to https://github.com/eclipse-gef/gef-classic/pull/778