Replace RunnableChain in DeferredUpdateManager with List - Bug 281327
In the class org.eclipse.draw2d.DeferredUpdateManager, Runnables are
queued using a special inner class RunnableChain which basically
produces a linked list of Runnables. When run is called on the head of
the linked list, first run of its next element is called an then run on
the runnable of the element itself.

This produces very large stacks, for each element in the list, a new
stack frame is created. Instead of the RunnableChain construct, a simple
ArrayList can be used as queue which will be faster and use less
memory.