Make FlyPaletteComposite contents use native double buffering
The FlyoutPaletteComposite is instantiated with style flag
SWT.DOUBLE_BUFFERED to use native double buffering. It's contained
TitleCanvas and ButtonCanvas do not use that flag. In consequence, their
LightweightSystem creates a BufferedGraphicsSource instead of a
NativeGraphicsSource for them.

This change adapts the TitleCanvas and ButtonCanvas instantiation to add
the SWT.DOUBLE_BUFFERED style, such that they use native double
buffering.