Handle HiDPI scaling in Zest Graph
This replicates the HiDPI scaling in GEF diagrams that was introduced
with 1bd28834343bce2bf73cbf66d9cb2accc54382f0, albeit in a simplified
form. Zest doesn't have an EditPart viewer and instead, the root layer
is directly embedded into the Graph widget.

When the native auto-scaling provided by SWT is disabled for the Graph,
a zoom listener is added that automatically synchronizes the monitor
zoom of the widget with the scale of the root figure.