Enable Draw2D auto-scaling by default
This renames the "draw2d.disableAutoscale" environment variable
introduced with 1bd28834343bce2bf73cbf66d9cb2accc54382f0 to
"draw2d.enableAutoscale". In addition, this property now needs to be
explicitly disabled.

By doing so, we drastically increase the visibility of this feature and
reduce the changes that need to be done by consumer projects.