Simplify creation of PopupHelper LWS when Draw2D autoscaling is enabled
When the auto-scaling property is set, a ScalableLayeredPane is injected
between the root figure and its contents.

The issue with this approach is that the contents is also stored as a
local variable of the LightweightSystem class, which is no longer
updated and therefore always null.

This pane is now embedded in a custom RootFigure subclass.