Override getChildren() in ScalableRootFigure of PopUpHelper
This is a follow-up to 4b38172a5fc7aff628a73eb9172cff7aa6c38d78, where
a custom RootFigure class was created, which injects a
ScalableLayeredPane between itself and its content.

The add() method was already overridden to add any additional children
to the scalable figure. One also needs to override the getChildren()
method to delegate to the same figure.