Only configure auto-scaling for the canvas when supported by SWT
The auto-scaling support in Draw2D relies on the existence of the
SHELL_ZOOM key, in order to calculate the zoom factor of the widget,
which was only added with the 2025-12 SWT release. Because auto-scaling
is enabled by default, Draw2D always paints figures at 100% zoom when
using an older Eclipse version, even if the monitor has a higher zoom
level.

By checking whether this key is set, this feature is only available with
the latest Eclipse release. With older versions, the native scaling
functionality provided by SWT is used.

Closes https://github.com/eclipse-gef/gef-classic/issues/812