Combine ScalableRootFigure and ScalableLayeredPane #818
This creates an internal ScalableLightweightSystem subclasses that is
used in the PopUpHelper. The root figure of this class implements the
IScalablePane interface and is automatically scaled to match the monitor
zoom.

This replaces the previous implementation, where a ScalableLayeredPane
was added between the root figure and its contents. The problem here is
that this pane is invisible to the layout manager and therefore the
figure, and all its children, have an empty size.

Rather than copy-pasting the same logic that is used in both the
ScalableFreeformLayeredPane and the ScalableLayered pane, the code is
instead moved to the IScalablePaneHelper. The subclasses still have to
override the relevant methods, but at least the implementation is now
centralized.

Closes https://github.com/eclipse-gef/gef-classic/issues/818