Added AutoscaleFreeformViewport to better encapsulate autoscaling
This converts the innerLayers of the FreeformGraphicalRootEditPart back
to a FreeformLayeredPane. In order to continue the support for the
Draw2D-based monitor-scaling, a custom AutoscaleFreeformViewport is
used, which is the new figure containing the
ScalableFreeformLayeredPane.

The goal of this change is to avoid visual bugs that may appear when
editors directly modify this innerLayer fields, as seen in:
https://github.com/eclipse-gef/gef-classic/pull/819

Co-authored-by: Patrick Ziegler <ziegler.patrick@hotmail.de>