Move configuration of auto-scaling mode to GraphicalViewerImpl
The SWT-based auto-scaling must only be disabled for viewers that are
based on IFigure's and can't therefore be done in the
AbstractEditPartViewer.

Otherwise the Outline view, which is populated with a TreeViewer, is
always drawn at 100% zoom, as there is no ScalableFigure to scale its
content to the monitor zoom.

Regression from 1bd28834343bce2bf73cbf66d9cb2accc54382f0