Dispose GC in FigureUtilities if Shell is destroyed #831
If the invisible Shell created by the FigureUtilities class is disposed,
any successive calls to GC.dispose() fail with a "Widget is disposed"
exception.

To avoid this, the checks in getShell() and setFonts() are hardened to
make sure that the previously created Shell/GC instances are still
valid.

For the Shell, dispose listener is added that explicitly invalidates any
previously created GC object.

Closes https://github.com/eclipse-gef/gef-classic/issues/831