Enable advanced mode for FigureUtilities GC
The GC used in FigureUtilities for text extent calculation does
currently not have advanced mode enabled. In most usage scenarios,
however, the context that uses the extents to render text will have a GC
with advanced mode enabled, as many GC operations such as applying
transforms and several draw operations implicitly activate it. As a
result, the extents calculated by the FigureUtilities will, in most
cases, not return appropriate results.

This change adapts the FigureUtilities to always use a GC in advanced
mode for extent calculation. It adapts few tests to conform to the
different widths returned for some of the used strings.

Fixes https://github.com/eclipse-gef/gef-classic/issues/809