Revert GC recreation in FigureUtilities
With commit efc73d2d6c2c076ad38190de2b8c0908a0537d18, GC disposals and
recreations were introduced in FigureUtilities. They used to fix a
memory leak caused by a long-living GC, as the operations executed on
such a GC were permanently stored in the GC instance (on Windows). That
change introduced two issues:
- In case a subclass of FigureUtilities performs further modifications
to the used GC, those modifications will be lost upon next call of
FigureUtilities#setFont(). This may be considered a breaking API change.
While creating such subclasses may be questionable, it's also the only
reason for keeping a static GC at all instead of using auto-disposed GCs
for the duration of the individual method calls.
- Resource leak tests became more difficult, as every test for a diagram
will usually leave behind a new GC (handle), precisely the last GC
created by the FigureUtilities upon setting a font during diagram usage.
That GC is detected as a leak and must somehow be excluded from leak
detected, such as using Sleak.

With a recent improvement to SWT, the GC implementation on Windows is
not prone to a memory leak when using a long-living GC on which only a
single operation (in this case setFont()) is performed repeatedly
anymore. Due to the mentioned issue and because the fix became
unnecessary with SWT enhancements, this reverts the previous change to
FigureUtilties.