Harmonize calculation of int-width and int-height in PrecisionRectangle
This is a follow-up to 116c3dd2d5c73db11d2d3702820dc881d274a6b8, due to
which the scale() method was overriden, similar to the performScale()
method.

When scaling a PrecisionRectangle, we need to make sure that its
int-representation is identical to a plain Rectangle, which be off by
one pixel compared to its double-representation.