Create a provisional PrecisionPointList to better handle rounding errors
When translating a PointList through multiple scalable layers, rounding
errors accumulate over time because the remainder is discarded after
each layer. For other shapes such as Points and Rectangles, this is
solved by using a "Precision" variant.

With this change, such a variant now also exists for the PointList and
is used when calling "translateToAbsolute()" or "translateToRelative()"
on a Figure.

Note: The implementation of this PrecisionPointList is not complete and
currently only works for this specific use case. The double-array
doesn't update itself when the int-array is modified. This class should
therefore not yet be used outside Draw2D.

Relates to https://github.com/eclipse-gef/gef-classic/issues/829