Update JRuby to 9.4.14.0
When verifying SSL certificates, jruby-openssl is not verifying that the
hostname presented in the certificate matches the one we are trying to
connect to, meaning a MITM could just present any valid cert for a
completely different domain they own, and JRuby wouldn't complain.

While this likely isn't relevant for the AsciiDoc generation, it is
still flagged by GitHub as potential vulnerability.