Explicitly set workflow permissions
If a GitHub Actions job or workflow has no explicit permissions set,
then the repository permissions are used. Repositories created under
organizations inherit the organization permissions.

Often these permissions do not adhere to the principle of least
privilege and can be reduced to read-only, leaving the write permission
only to a specific types as issues: write or pull-requests: write.