Do not wrap custom Translatable representations as precise types
The wrapping mechanism for Translatables when performing
Figure#translate...() operations currently wraps every instance of the
Translatable types provided by Draw2d. However, consumers may have
custom specializations of those types, which can be incompatible with
the implemented wrapping. This particularly applies to types like
PointList, for which precision implementations did not exist such that
consumers may use custom implementations of them.

This adapts the wrapping logic to ensure that only the relevant basic
Draw2d types are wrapped to avoid that custom specializations that are
potentially incompatible with the implemented wrapping mechanism get
wrapped as well.