Restore PrecisionRectangle substitutability for Rectangle
Recent changes to PrecisionRectangle adapted how the integer-based
size values are calculated, which introduced inconsistencies with other
operations and data (such as the relation between bottom-right and
top-left corner of a rectangle) and which may also break existing
consumers due to the changed behavior of API methods.

This change reverts the changes to the necessary extent such that the
existing behavior is restored. It also adds tests for some of the
scenarios that broke with the changes.