Make IFigure#translate...() results backward compatible for Rectangle
The wrapping of Rectangles into their precision representations when
calling IFigure#translate...() methods leads to results that are
incompatible with pre-existing behavior when only a single scaled layer
is involved. The reason is the different rounding applied in a basic
Rectangle compared to a PrecisionRectangle. Recent attempts to adapt the
PrecisionRectangle behavior failed because they broke backward
compatibility. But without streamlining the rounding of Rectangle and
PrecisionRectangle, the results of IFigure#translate...() become
incompatible.

For that reason, this change adapts the unwrapping of PrecisionRectangle
to Rectangle in IFigure#translate...() to use the same rounding as
applied in Rectangle to ensure backward compatibility of the method's
behavior.