Only use double-precision translation when HighDPI scaling is enabled
When the "draw2d.enableAutoscale" system property is enabled, additional
(scalable) layers are injected into the IFigure hierarchy to take the
monitor zoom into consideration.

As a result of this, translating geometric shapes may cause additional
rounding errors, which are the result of translating integer-values with
an additional, fractional zoom level.

To work around this problem, the "useDoublePrecision()" method was added
with 5cfe5b7587123e68c2077b6dbec4e5657316d458, which converts the
geometric shapes to their precise variants. This conversion is generally
not backwards compatible.

To make sure clients can fully opt-out of this new behavior, this
conversion should only be done when the "draw2d.enableAutoscale" is set
to true, meaning only when these new layers are added, that require this
extended translation logic.