Embed tracker for dragging flyout palette composite into control
The tracker for dragging the flyout palette is currently instantiated on
the display rather than the control itself. This leads to offsets with
monitor-specific scaling on Windows, as the display is lacking the
monitor-specific zoom context.

This change adapts the tracker to operate on the flyout palette control
rather than the display. It simplifies the implementation as the
control/display coordinate conversions can be removed.

Fixes https://github.com/eclipse-gef/gef-classic/issues/868