Correct mapping of tracker size for FlyoutPaletteComposite movement
When dragging a flyout palette, a tracker on the Display is instantiated
to show a preview of the drag result. When different zooms are involved
(such as when using Draw2d autoscaling for the monitor-specific scaling
capability on Windows), the tracker preview has a wrong size on non-100%
monitors. This is caused by multiple Display#map() methods being applied
on the same rectangle, which leads to the zoom being applied multiple
times.

This change adapts the preview size calculation to prevent that the map
operation (and thus the zoom) is not applied multiple times to the width
and height of the preview rectangle. The change is backward compatible
as without monitor-specific scaling capabilities and Draw2d autoscaling
there are no zoom transformations involved, which is why the multiple
map operations applied did not lead to any issues.

Fixes https://github.com/eclipse-gef/gef-classic/issues/868