Create custom cursor for 100%, 150% and 200% zoom #872
At least on Windows and Linux, the operating system doesn't seem to
linearly scale the system cursor.

Instead, predefined cursor images are used for:

- ( ..., 150%) - Cursor at 100% monitor zoom
- [150%, 200%) - Cursor at 150% monitor zoom
- (200%,  ...) - Cursor at 200% monitor zoom

SWT is linearly scaling the cursor icon. So if at e.g. 125% zoom, the
cursor inside the widget will be scaled to 125% zoom but is not scaled
outside the widget.

To account for this, the SWT-based cursor is replaced with three cursor
SVGs, one for each of those zoom levels. Note that the images are split
up to make sure the line width is exactly one for each cursor.

Closes https://github.com/eclipse-gef/gef-classic/issues/872