Implement remaining methods of PrecisionPointList
This overrides all methods in the PointList that require
double-precision handling.

Unlike the other precision classes, the double-array in the
PrecisionPointList only keeps track of the fractional parts of all
points. This has the advantage that we don't need to

Instead of keeping a double-precision copy of the (x,y) coordinates of
all contained points, the PrecisionPointList only keeps the decimal
parts. This safes us the overhead of keeping both arrays in sync, at the
cost of a minor overhead when accessing the precise points.

Closes https://github.com/eclipse-gef/gef-classic/issues/883