Prepare deprecated fields in geometry classes for internalization
The public fields of the PrecisionPoint, PrecisionRectangle and
PrecisionDimension have been annotated with @noreference with
7865d7b918e456cb110405030e31e059ea045f38 and deprecated with
390e597e7523d4739154be895934b19997931f0f, over 15 years ago.

Given the amount of times those fields have been public and the
frequency with which they might be accessed by consumers, the removal
will be split into two parts:

1) New private fields are added to the geometry classes that track the
fractional part of the x, y, width and height. Calling any of the
getters no longer returns the public field but instead calculates the
precise value as follows:

- preciseX() -> x + fractX
- preciseY() -> y + fractY
- preciseWidth() -> width + fractWidth
- preciseHeight() -> height + fractHeight

2) The following fields/methods are removed:
- preciseX
- preciseY
- preciseWidth
- preciseHeight
- updateInts()

Consumers can try out the "new" behavior by migrating to the method
calls, but always fall back to the public fields in case of
incompatibilities.

Note: There is one thing that can't be replicated once the public fields
have been removed; If the integer-field is updated, the fractional part
is cleared when calling one of the methods. If for example the integer
part of the value 2.5 is changed to 3, the new value will now be 3.5 as
opposed to 3.