Replace usage of XYScaledGraphics with a plain (SWT)Graphics instance
The XYScaledGraphics class has been created as a subclass of
ScaledGraphics, in order to independently scale along the x-axis and the
y-axis.

It is used inside the GraphContainer to shrink its members based on the
width and height of the container, to make them smaller as the nodes in
their parent container.

The same behavior can be achieved when using an SWTGraphics, which also
avoids the problems that arise with the scaling done by ScaledGraphics.

Closes https://github.com/eclipse-gef/gef-classic/issues/950