Catch ServiceConfigurationError when JSVG is not in classpath
When Draw2D is used in a pure Java environment, it might happen that the
JSVG library, which provides the SVG support, is not in the classpath.

The call to isSvgSupported() should return `false` in such a situation,
rather than throwing a ServiceConfigurationError. This problem can be
observed when e.g. running any of the widgets that access an SVG.