Allow signing builds from the maintenance branch
The Jenkinsfile currently only allows signing and therefore also
deploying the artifacts from the main branch to our update site. In
order to allow the same for our maintenance branch, this rule needs to
be relaxed.

Instead of requiring the branch to have the name 'master', it may also
have the form 'Rx_yy_maintenance', where 'x' and 'y' are integers.