Use correct generic types in EditPartViewer #903
The generic type of the following methods has to be adapted, in order to
account for the fact that an EditPartViewer might be based on both
IFigure's (GraphicalEdiPartViewer) and Widget's (TreeViewer):

- Map<~~IFigure~~?, EditPart> getVisualPartMap()`
- findObjectAtExcluding(Point, Collection<~~IFigure~~?>)
- findObjectAtExcluding(Point, Collection<~~IFigure~~?>, Conditional)

To reduce the impact this has on downstream projects, the
`getVisualPartMap()` method is overridden by both the graphical viewer
and the tree viewer, where the returned map is casted to its expected
type.

For the `findObjectAtExcluding()`, the `IFigure` is only used in the
exclusion set for graphical viewers. For tree viewers, the `TreeEdiPart`
is used directly.

Closes https://github.com/eclipse-gef/gef-classic/issues/903