Make the creation of the PopUpHelper configurable
This introduces a new `IToolTipHelperFactory` that is used as an OSGi
service in the `DomainEventDispatcher` to allow consumers to inject
their own `ToolTipHelper`.

It is currently very difficult to customize the tool-tips shown by an
edit-part viewer in any meaningful ways, as one has to:

a) Subclass the event dispatcher to use a custom tool-tip helper
b) Subclass the LWS to use this custom event dispatcher
c) Subclass the viewer to use this custom LWS

A similar customization was already done in the past via
e9839e7588032a97ede7c1fbdb8503d7a0b908aa. The advantage of
using a service is that it allows implementations to customize
the tool-tip of a specific viewer, rather than all viewers.