Further refinement of the return type of getVisualPartMap() #903
This is a continuation of c5b0306ff812205be4d9dd6b554c476b0fd61095. Due
to this change, consumers may now experience a compilation error when
calling `getViewer().getVisualPartMap().put(...)`.

When calling `getVisualPartMap()`, a wildcard generic map is returned,
which does not allow the addition of values. To get around this, one has
to cast the map to a specific type, which is not something that should
be done on consumer side. But the keys also can't be plain `Object`s,
because this prevents overriding this method in sub-classes.

To get around this limitation, the following approach is taken: The
signature in the `EditPartViewer` interface is changed so that this
method returns a generic map. The generic type is not specified, so
calling `getVisualPartMap()` returns a `Map<Object,EditPart`, which
allows the addition of entries.

But when the `GraphicalEditPartViewer` and `TreeViewer`, this method can
be overridden to return a Map of `IFigure`'s or `Widget`s, respectively.