Generify `extendedData` + accessors in Request #155
The extended data of a request contains arbitrary data. Because the map
is supposed to be mutable, the getter has to return a `Map<Object,
Object>` and because the request should accept any Map, wildcards have
to be used. The unchecked typecast from a `Map<?,?>` to
`Map<Object,Object>` is generally safe, as every type is also an
`Object`.

Contributes to https://github.com/eclipse-gef/gef-classic/issues/155