Generify EditPolicyIterator constructor #155
This iterator expects a list of key+EditPolicy "pairs". To remain
backwards compatible, a wildcard should be used so that both a
`List<Object>` and `List<?>` are valid arguments.

Because the iterator should not be instantiated by subclasses, the
already deprecated constructor is now also marked as "for removal".

Contributes to https://github.com/eclipse-gef/gef-classic/issues/155