Generify PaletteViewerKeyHandler.getNavigationSiblings() #155
The generic type is already defined by the PaletteViewerKeyHandler. This
requires the `getParent()` method in the `PaletteEditPart` and the
`getFocusEditPart()` in the `PaletteViewerKeyHandler` to be overridden,
as `EditPart`s are generally not required to return a
`GraphicalEditPart`. However, inside a palette only `PaletteEditPart`s
should be used.

Contributes to https://github.com/eclipse-gef/gef-classic/issues/155