Use higher-resolution image data in SWTGraphics if available #922
When drawing an image on a scaled `Graphics` object, the image data at
100% zoom is taken and then artificially up-scaled. This results in
blurry images when scaling by high factors.

This change adapts the `FileImageDataProvider` and the `SWTGraphics` to
take this problem into consideration, especially when dealing with SVGs.

The `FileImageDataProvider` is currently not able to return the
`ImageData` at arbitrary zoom levels, as this is not supported by SWT,
which requires the `ImageFileNameProvider` (or the internal
`ImageDataAtSizeProvider`) to be implemented. The problem with the
`ImageFileNameProvider` is that it can only return proper file paths. So
if an image is embedded inside a jar, it first needs to be extracted
into a temporary file.

The `SWTGraphics` then needs to use the new
`drawImage(Image,int,int,int,int)` method so that SWT uses this provider
to probler scale the image to the correct size.