Implement getAbsoluteScale for SWT Graphics
As SWT Graphics allows different scale factors in x and y, rotations,
and shear which is all stored in the transformation matrix it is not as
easy to get a single value for the current absolute scale. This
implementation derives an approximation from the determinate of the
matrix. This represents an average scale factor. For non rotated uniform
x/y scaling it is the exact scale factor.

Even if it is an approximation it is suitable for level of detail
checking code.

Implements: https://bugs.eclipse.org/bugs/show_bug.cgi?id=267462